[REF] core: separated odoo.orm.types
Define generic ORM types in a single place.
This is only to define typing and not used during execution.
