+++ odoo/orm/fields.py
@@ -43,7 +43,7 @@ from odoo.tools.misc import unquote, has_list_types, Sentinel, SENTINEL
 from odoo.tools.translate import html_translate
 
 from odoo.osv import expression
-from .api import ContextType, DomainType, IdType, NewId, ValuesType
+from .api import IdType, NewId
 from .utils import expand_ids, check_pg_name, check_property_field_value_name, PREFETCH_MAX
 # we may not import models here to avoid cyclic dependencies
 
@@ -51,6 +51,7 @@ T = typing.TypeVar("T")
 # BaseModel will be injected into this module by loading odoo.orm.models
 if typing.TYPE_CHECKING:
     from .models import BaseModel
+    from .types import ContextType, DomainType, ValuesType
     _Fields_BaseModel = BaseModel
     M = typing.TypeVar("M", bound=BaseModel)
 else:
